+++ packages/anon-aadhaar-react/src/index.ts
@@ -2,9 +2,11 @@ import { LogInWithAnonAadhaar } from './components/LogInWithAnonAadhaar'
 import { useAnonAadhaar } from './hooks/useAnonAadhaar'
 import { AnonAadhaarProvider } from './provider/AnonAadhaarProvider'
 import { AnonAadhaarProof } from './components/AnonAadhaarProof'
+import { LogInWithAnonAadhaarV2 } from './components/LoginV2/LogInWithAnonAadhaarV2'
 
 export {
   LogInWithAnonAadhaar,
+  LogInWithAnonAadhaarV2,
   useAnonAadhaar,
   AnonAadhaarProvider,
   AnonAadhaarProof,